requirements.txt
pytest==8.3.3
allure-pytest==2.13.5
pytest-html==4.1.1
Appium-Python-Client~=4.2.0
selenium~=4.25.0
PyYAML~=6.0.2
pytest-xdist==3.6.1
jdatetime